Migrate generated code tests out of `pkg`
As part of #1142, we want to reduce the transitive dependencies of
oapi-codegen, which requires creating a multi-module Go project.

Before we can do this, we need to break some dependencies between
packages that don't make sense.
